feat: 1. Kеширования на горячие маршруты 2. Cбор метрик (grafana/prometeus) 3. Документация swagger 4. нагрузочное тестирование

+++ internal/transport/middleware/context/getfromcontext.go
@@ -21,8 +21,8 @@ var (
 	ErrFailedIsAdminCheck = errors.New("failed to check if user is admin")
 )
 
-func APPIDFromContext(ctx context.Context) (int, bool) {
-	appid, ok := ctx.Value(AppIDKey).(int)
+func APPIDFromContext(ctx context.Context) (uint64, bool) {
+	appid, ok := ctx.Value(AppIDKey).(uint64)
 	return appid, ok
 }
